+++ app-phoenix/priv/repo/oban_migrations/20260904103515_install_oban.exs
@@ -1,0 +1,6 @@
+defmodule Dawarich.Repo.Migrations.InstallOban do
+  use Ecto.Migration
+
+  def up, do: Oban.Migration.up(prefix: "oban")
+  def down, do: Oban.Migration.down(prefix: "oban")
+end